Split e2e tests into base and files
Split test files further
