+++ tests/e2e/tests/utils.ts
@@ -1,39 +1,15 @@
-import { readFileSync } from 'fs';
 import * as got from 'got';
-import * as path from 'path';
 
 
-const dataDir = path.join(__dirname, '../../data');
-const dataPath = (name: string, ...dirs: string[]) =>
-  path.join(dataDir, ...dirs, name);
+const maxAppWarmup = 45;
+export const rootURL = process.env.SIGNHASH_URL || 'http://localhost:8080';
 
 
-export const readDataFile = (name: string) => readFileSync(dataPath(name));
-
-export const accounts = readDataFile('accounts.txt').toString()
-  .trim()
-  .split('\n')
-
-
-export interface TestFile {
-  path: string;
-  name: string;
-  checksum: string;
-  signers: number[];
+export function rootFixture(fixture: FixtureFn) {
+  return fixture
+    .page(rootURL)
+    .before(waitForPage(rootURL, maxAppWarmup));
 }
-
-
-export const buildTestFile =
-  (accounts: string[]) => (name: string): TestFile => {
-    const path = dataPath(name, 'files');
-    const meta = JSON.parse(readFileSync(path + '.json').toString());
-    return {
-      path,
-      name,
-      signers: meta.signers,
-      checksum: meta.checksum,
-    }
-  }
 
 
 export function waitForPage(url: string, timeout: number) {
